+++ Forge/build.gradle
@@ -30,10 +30,20 @@ compileJava {
 }
 jar {
     manifest {
-        attributes 'FMLCorePlugin': 'customskinloader.forge.ForgePlugin'
-        attributes 'FMLCorePluginContainsFMLMod': true
+        attributes([
+                "FMLCorePlugin"              : "customskinloader.forge.ForgePlugin",
+                "FMLCorePluginContainsFMLMod": true,
+                "Specification-Title"        : "CustomSkinLoader",
+                "Specification-Vendor"       : "xfl03",
+                "Specification-Version"      : "1",
+                "Implementation-Title"       : "CustomSkinLoader",
+                "Implementation-Version"     : "${version}",
+                "Implementation-Vendor"      : "xfl03",
+                "Implementation-Timestamp"   : new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")
+        ])
     }
     exclude 'net/minecraft/client/renderer/texture/NativeImage.class'
+    exclude 'net/minecraftforge/fml/common/Mod.class'
 }
 sourceSets {
     main {
@@ -76,10 +86,6 @@ processResources {
             gitVersion    : System.getenv("CIRCLE_SHA1"),
             buildUrl      : System.getenv("CIRCLE_BUILD_URL")
     ])}
-}
-
-dependencies {
-    //compileOnly "cpw.mods:modlauncher:0.1.0-rc.3"
 }
 
 task signJar(type: SignJar, dependsOn: reobfJar) {
