+++ Forge/resource/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+    "pack": {
+        "description": "CustomSkinLoader resources",
+        "pack_format": 4
+    }
+}
